[Modify] project structure
1. Modify rename `bpfocket.h` -> `bpfapture.h`
2. Modify `RawSocket` class -> `BPFapture`
3. Add master header file (include/bpfocket.h)
3. typo
