+++ test/CMakeLists.txt
@@ -5,8 +5,8 @@
 message("FetchContent: googletest")
 include(FetchContent)
 FetchContent_Declare(googletest
-  GIT_REPOSITORY "https://github.com/google/googletest.git"
-  GIT_TAG        "v1.14.0"
+    GIT_REPOSITORY "https://github.com/google/googletest.git"
+    GIT_TAG        "v1.14.0"
 )
 # For Windows: Prevent overriding the parent project's compiler/linker settings
 set(gtest_force_shared_crt ON CACHE BOOL "" FORCE)
